Redesign logging manipulators
