+++ libs/logging/CMakeLists.txt
@@ -11,29 +11,21 @@ list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 # Default location is $HPX_ROOT/libs/logging/include
 set(logging_headers
   hpx/logging.hpp
-  hpx/logging/detail/time_format_holder.hpp
   hpx/logging/detail/macros.hpp
-  hpx/logging/detail/manipulator.hpp
   hpx/logging/detail/logger.hpp
-  hpx/logging/detail/format_write_detail.hpp
   hpx/logging/detail/cache_before_init.hpp
   hpx/logging/format.hpp
-  hpx/logging/format/array.hpp
-  hpx/logging/format/formatter/named_spacer.hpp
   hpx/logging/format/formatter/high_precision_time.hpp
   hpx/logging/format/formatter/defaults.hpp
   hpx/logging/format/formatter/thread_id.hpp
   hpx/logging/format/named_write.hpp
-  hpx/logging/format/destination/named.hpp
   hpx/logging/format/destination/defaults.hpp
   hpx/logging/format/destination/file.hpp
   hpx/logging/level.hpp
   hpx/logging/logging.hpp
+  hpx/logging/manipulator.hpp
   hpx/logging/message.hpp
-  hpx/logging/writer/format_write.hpp
-  hpx/logging/writer/named_write.hpp
 )
-
 
 # Default location is $HPX_ROOT/libs/logging/include_compatibility
 set(logging_compat_headers
@@ -41,24 +33,18 @@ set(logging_compat_headers
   hpx/util/logging/format.hpp
   hpx/util/logging/writer/format_write.hpp
   hpx/util/logging/writer/named_write.hpp
-  hpx/util/logging/detail/time_format_holder.hpp
   hpx/util/logging/detail/macros.hpp
   hpx/util/logging/detail/level.hpp
   hpx/util/logging/detail/manipulator.hpp
   hpx/util/logging/detail/logger.hpp
-  hpx/util/logging/detail/format_write_detail.hpp
   hpx/util/logging/detail/cache_before_init.hpp
   hpx/util/logging/logging.hpp
-  hpx/util/logging/format/formatter/named_spacer.hpp
   hpx/util/logging/format/formatter/high_precision_time.hpp
   hpx/util/logging/format/formatter/defaults.hpp
   hpx/util/logging/format/formatter/thread_id.hpp
-  hpx/util/logging/format/array.hpp
   hpx/util/logging/format/named_write.hpp
-  hpx/util/logging/format/destination/named.hpp
   hpx/util/logging/format/destination/defaults.hpp
   hpx/util/logging/format/destination/file.hpp
-  hpx/util/logging/format/op_equal.hpp
   hpx/util/logging/format/optimize.hpp
 )
 
@@ -66,7 +52,9 @@ set(logging_compat_headers
 set(logging_sources
   level.cpp
   logging.cpp
+  manipulator.cpp
   message.cpp
+  format/named_write.cpp
 )
 
 include(HPX_AddModule)
