+++ libs/logging/include_compatibility/hpx/util/logging/detail/manipulator.hpp
@@ -6,16 +6,16 @@
 
 #include <hpx/config.hpp>
 #include <hpx/logging/config/defines.hpp>
-#include <hpx/logging/detail/manipulator.hpp>
+#include <hpx/logging/manipulator.hpp>
 
 #if defined(HPX_LOGGING_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/util/logging/detail/manipulator.hpp is deprecated,         \
-    please include hpx/logging/detail/manipulator.hpp instead")
+    please include hpx/logging/manipulator.hpp instead")
 #else
 #warning                                                                       \
     "The header hpx/util/logging/detail/manipulator.hpp is deprecated,    \
-    please include hpx/logging/detail/manipulator.hpp instead"
+    please include hpx/logging/manipulator.hpp instead"
 #endif
 #endif
