+++ libs/logging/include_compatibility/hpx/util/logging/writer/named_write.hpp
@@ -6,15 +6,15 @@
 
 #include <hpx/config.hpp>
 #include <hpx/logging/config/defines.hpp>
-#include <hpx/logging/writer/named_write.hpp>
+#include <hpx/logging/named_write.hpp>
 
 #if defined(HPX_LOGGING_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/util/logging/writer/named_write.hpp is deprecated, \
-    please include hpx/logging/writer/named_write.hpp instead")
+    please include hpx/logging/named_write.hpp instead")
 #else
 #warning "The header hpx/util/logging/writer/named_write.hpp is deprecated, \
-    please include hpx/logging/writer/named_write.hpp instead"
+    please include hpx/logging/named_write.hpp instead"
 #endif
 #endif
