+++ libs/logging/src/logging.cpp
@@ -22,12 +22,6 @@
 #include <utility>
 #include <vector>
 
-#if defined(HPX_MSVC_WARNING_PRAGMA)
-#pragma warning(push)
-// 'class1' : inherits 'class2::member' via dominance
-#pragma warning(disable : 4250)
-#endif
-
 ///////////////////////////////////////////////////////////////////////////////
 namespace hpx { namespace util {
     HPX_DEFINE_LOG(agas, disable_all)
@@ -81,7 +75,7 @@ namespace hpx { namespace util {
             }
         }
     }    // namespace detail
-}}    // namespace hpx::util
+}}       // namespace hpx::util
 
 ///////////////////////////////////////////////////////////////////////////////
 #include <hpx/logging/detail/cache_before_init.hpp>
@@ -104,115 +98,5 @@ namespace hpx { namespace util { namespace logging { namespace detail {
         }
     }
 }}}}    // namespace hpx::util::logging::detail
-
-///////////////////////////////////////////////////////////////////////////////
-#include <hpx/logging/format/destination/file.hpp>
-
-namespace hpx { namespace util { namespace logging { namespace destination {
-    file::mutex_type file::mtx_ = BOOST_DETAIL_SPINLOCK_INIT;
-}}}}    // namespace hpx::util::logging::destination
-
-///////////////////////////////////////////////////////////////////////////////
-#include <hpx/logging/format/destination/named.hpp>
-
-namespace hpx { namespace util { namespace logging { namespace destination {
-    namespace detail {
-        void named_context::compute_write_steps()
-        {
-            m_info.write_steps.clear();
-
-            std::istringstream in(m_info.format_string);
-            std::string word;
-            while (in >> word)
-            {
-                if (word[0] == '+')
-                    word.erase(word.begin());
-                else if (word[0] == '-')
-                    // ignore this word
-                    continue;
-
-                if (m_info.name_to_destination.find(word) !=
-                    m_info.name_to_destination.end())
-                    m_info.write_steps.push_back(
-                        m_info.name_to_destination.find(word)->second);
-            }
-        }
-}}}}}    // namespace hpx::util::logging::destination::detail
-
-///////////////////////////////////////////////////////////////////////////////
-#include <hpx/logging/format/formatter/named_spacer.hpp>
-
-namespace hpx { namespace util { namespace logging { namespace formatter {
-    namespace detail {
-        static std::string unescape(std::string escaped)
-        {
-            typedef std::size_t size_type;
-            size_type idx_start = 0;
-            while (true)
-            {
-                size_type found = escaped.find("%%", idx_start);
-                if (found != std::string::npos)
-                {
-                    escaped.erase(
-                        escaped.begin() + static_cast<std::ptrdiff_t>(found));
-                    ++idx_start;
-                }
-                else
-                    break;
-            }
-            return escaped;
-        }
-
-        void base_named_spacer_context::compute_write_steps()
-        {
-            typedef std::size_t size_type;
-
-            m_info.write_steps.clear();
-            std::string remaining = m_info.format_string;
-            size_type start_search_idx = 0;
-            while (!remaining.empty())
-            {
-                size_type idx = remaining.find('%', start_search_idx);
-                if (idx != std::string::npos)
-                {
-                    // see if just escaped
-                    if ((idx < remaining.size() - 1) &&
-                        remaining[idx + 1] == '%')
-                    {
-                        // we found an escaped char
-                        start_search_idx = idx + 2;
-                        continue;
-                    }
-
-                    // up to here, this is a spacer string
-                    start_search_idx = 0;
-                    std::string spacer = unescape(remaining.substr(0, idx));
-                    remaining = remaining.substr(idx + 1);
-                    // find end of formatter name
-                    idx = remaining.find('%');
-                    format_base_type* fmt = nullptr;
-                    if (idx != std::string::npos)
-                    {
-                        std::string name = remaining.substr(0, idx);
-                        remaining = remaining.substr(idx + 1);
-                        fmt = m_info.name_to_formatter[name];
-                    }
-                    // note: fmt could be null, in case
-                    m_info.write_steps.push_back(write_step(spacer, fmt));
-                }
-                else
-                {
-                    // last part
-                    m_info.write_steps.push_back(
-                        write_step(unescape(remaining), nullptr));
-                    remaining.clear();
-                }
-            }
-        }
-}}}}}    // namespace hpx::util::logging::formatter::detail
-
-#if defined(HPX_MSVC_WARNING_PRAGMA)
-#pragma warning(pop)
-#endif
 
 #endif    // HPX_HAVE_LOGGING
